# person api 관련 controller, service 등 객체 작성 및 리팩토링

+++ jpa/src/main/kotlin/com/comento/jpa/domain/person/PersonRepository.kt
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PersonRepository: CrudRepository<Person, Int> {
-    fun findByPersonId(personId: Int): Person
+interface PersonRepository: CrudRepository<Person, Long> {
+    fun findByPersonId(personId: Long): Person
 }
