+++ jpa/src/main/kotlin/com/comento/jpa/presentation/PersonController.kt
@@ -1,12 +1,8 @@
 package com.comento.jpa.presentation
 
-import com.comento.jpa.domain.person.Person
 import com.comento.jpa.domain.person.dto.PersonDto
 import com.comento.jpa.domain.person.dto.PersonResponse
-import com.comento.jpa.domain.person.dto.PersonSaveDto
-import com.comento.jpa.domain.person.dto.PersonUpdateDto
 import com.comento.jpa.service.PersonService
-import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
@@ -24,7 +20,7 @@ class PersonController(
     fun findAll() = service.findAll()
 
     @PostMapping("/{personId}")
-    fun findById(@PathVariable personId: Int) = service.findByPersonId(personId)
+    fun findById(@PathVariable personId: Long) = service.findByPersonId(personId)
 
     @PostMapping
     fun save(@RequestBody personDtoList: List<PersonDto>): PersonResponse = service.save(personDtoList)
@@ -34,7 +30,4 @@ class PersonController(
 
 //    @DeleteMapping
 //    fun delete()
-
-
-
 }
